various: add view role to admins
OpenShift will not be happy with users that try to look for their own
projects/namespaces or use `oc project` if they do not have at least
this permission in those namespaces. While the previous roles are
technically the minimal set, this additional role does not give them any
privileges that are worrisome and will allow the CLI to operate clearly.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>