content/en/docs/how-tos/onboarding-a-new-component: Update ART link
Mojo has been deprecated, and the clone-able ART ticket I'm linking
instead opens with:

> First read Guidelines for requesting new content managed by OCP ART
> (at least "Getting Started" section). There is much to be done
> before you can use this template.

which links the successor article to the old Mojo page.  I'm also
dropping the dist-git warning, because if we need to underline things
like that within the ART steps, the most central place for the
underlining is the ART instructions themselves.