content/en/docs/architecture/ci-operator-internals: List namespace timeouts
So folks know how long they need to wait before kicking off a retest
if they want to pick up a fresh namespace with new release-image
content.  As an alternative to inlining the values here, we could
expose them in CI-operator logging, and then tell folks which log
lines to look for, or some such.  As long as the process is less
mysterious.