content/en/docs/architecture/ci-operator-internals/observer-pods: Wait-for-KUBECONFIG
It seems unlikely that observers would do something in the presence of
kubeconfig, but be fine skipping that thing in its absence.  Pivot the
example to a poll loop that will wait until a kubeconfig is available.
Personally I'd like gathered step logs to use --timestamps, but in the
absence of that, I'm manually including timestamps in my output lines
to make it easier to tell when the kubeconfig shows up.