how-tos: use better cluster profile example
The pull request currently linked does not contain the Boskos `reaper`
configuration changes which should be added for new resources.  Even though the
link right next to it describes the requirements, resources are still added
without the corresponding configuration.

Replace the link with one to a pull request containing all required changes.