Move "internals" to its own section
This started as a small section on `ci-operator` but is now composed of several
pages and covers other services.  It also now occupies a good portion of the
side menu, even though only Test Platform members will likely be interested in
the contents.

Moving to its own top-level section solves both problems.