content/en/docs/how-tos/onboarding-a-new-component: Self-checks for image presence
Supplement "sync with ART" with some steps that all OpenShift admins
can use to self-serve "yes, my image $NAME is in position for both ART
and CI $VERSION ImageStreams".

Also provide an example of "I'm not quite sure what my image's name is
in this ImageStream..." lookup.

I'd tried linking to [1], so folks can look for builds to see if their
image is failing to build.  But the ART dashboard requires VPN access,
and the links presubmit was failing [2]:

  INFO[2023-10-31T20:15:23Z] Using html-proofer version: 3.19.0
  With config options https://github.com/gjtorikian/html-proofer/blob/v3.19.0/lib/html-proofer/configuration.rb
  Running ["LinkCheck", "ImageCheck", "ScriptCheck"] on ["."] on *.html...
  Checking 435 external links...
  Ran on 64 files!
  - ./public/docs/how-tos/onboarding-a-new-component/index.html
    *  External link https://art-dash.engineering.redhat.com/dashboard/build/history failed: response code 0 means something's wrong.
               It's possible libcurl couldn't connect to the server or perhaps the request timed out.
               Sometimes, making too many requests at once also breaks things.
               Either way, the return message (if any) from the server is: Couldn't resolve host name
  HTML-Proofer found 1 failure!

so I'm not linking the ART dashboard.

[1]: https://art-dash.engineering.redhat.com/dashboard/build/history
[2]: https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_ci-docs/399/pull-ci-openshift-ci-docs-master-links/1719445752641490944