jira: update premerge verification documentation
Add documentation for new `/verified by` and `/verified later` commands
for the jira-lifecycle-plugin.

This PR also removes documentation for an older premerge verification
method that was added for QE. This method ended up being unused, so the
documentation has been removed to avoid confusion with the new premerge
verification method.